+++ libs/rust/sdk/src/actions/impls/mod.rs
@@ -1,5 +1,10 @@
+use std::collections::HashMap;
+use std::sync::Arc;
+
 use slog::debug;
 
+use crate::actions::Action;
+use crate::actions::ActionHook;
 use crate::Agent;
 use crate::AgentContext;
 
@@ -8,9 +13,14 @@ pub(crate) mod debug;
 mod service;
 
 /// Register standard agent actions.
-pub fn register_std_actions(agent: &dyn Agent, context: &AgentContext) {
+pub fn register_std_actions(
+    agent: &dyn Agent,
+    context: &AgentContext,
+    hooks: HashMap<ActionHook, Arc<dyn Action>>,
+) {
     debug!(context.logger, "Registering standard actions");
-    self::service::register(agent, context);
+    let graceful = hooks.get(&ActionHook::StoreGracefulStop).cloned();
+    self::service::register(agent, context, graceful);
 
     #[cfg(any(debug_assertions, test))]
     self::debug::register_debug_actions(context);