+++ libs/rust/sdk/src/actions/impls/service/mod.rs
@@ -1,6 +1,9 @@
+use std::sync::Arc;
+
 use serde::Deserialize;
 use serde::Serialize;
 
+use crate::actions::Action;
 use crate::actions::ACTIONS;
 use crate::Agent;
 use crate::AgentContext;
@@ -35,10 +38,10 @@ impl Default for ServiceActionState {
 }
 
 /// Register all service related actions.
-pub fn register(agent: &dyn Agent, context: &AgentContext) {
+pub fn register(agent: &dyn Agent, context: &AgentContext, graceful: Option<Arc<dyn Action>>) {
     let supervisor = self::supervisor::factory(agent, context);
-    ACTIONS::register_reserved(GracefulRestart::make(agent, &supervisor));
-    ACTIONS::register_reserved(GracefulStop::make(agent, &supervisor));
+    ACTIONS::register_reserved(GracefulRestart::make(graceful.clone(), &supervisor));
+    ACTIONS::register_reserved(GracefulStop::make(graceful, &supervisor));
     ACTIONS::register_reserved(ServiceRestart::make(&supervisor));
     ACTIONS::register_reserved(ServiceStart::new(&supervisor));
     ACTIONS::register_reserved(ServiceStop::new(&supervisor));