+++ libs/rust/sdk/CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Optional `ActionHook`s to provide store-specific implementation of SDK actions.
+
 ### Changed
+- **BREAKING**: Removed Agent trait method to expose the graceful stop action.
 - **BREAKING**: Rework action kinds format.
 
 ## [0.4.1] - 2020-03-07